重新优化用户模块

+++ src/main/java/estate/dao/impl/UserDaoImpl.java
@@ -1,5 +1,6 @@
 package estate.dao.impl;
 
+import estate.common.config.AppUserStatus;
 import estate.common.config.UserType;
 import estate.dao.UserDao;
 import estate.entity.database.AppUserEntity;
@@ -33,40 +34,6 @@ public class UserDaoImpl implements UserDao
 
 
 
-    public AppUserEntity getUserByPhone(String phone)
-    {
-        System.out.print("dsfds");
-        return null;
-    }
-
-    @Override
-    public AppUserEntity getByPhoneStatus(String phone, Byte status)
-    {
-        Session session=getSession();
-        String hql;
-        List list;
-        if (status!=null)
-        {
-            hql = "from AppUserEntity t where t.phone=:phone and t.status=:status";
-            list = session.createQuery(hql).setString("phone", phone).setByte("status", status).list();
-        }
-        else
-        {
-            hql = "from AppUserEntity t where t.phone=:phone";
-            list= session.createQuery(hql).setString("phone", phone).list();
-        }
-        if (list.size()>0)
-            return (AppUserEntity) list.get(0);
-        return null;
-    }
-
-    @Test
-    public void test ()
-    {
-        this.getUserByPhone("18122392");
-        System.out.print("dsfds");
-    }
-
     public TableData getOwnerList(TableFilter tableFilter)
     {
         Session session = getSession();
@@ -88,71 +55,9 @@ public class UserDaoImpl implements UserDao
 
         tableData.setRecordsFiltered(count);
         tableData.setJsonString(list);
-        tableData.setRecordsTotal(this.count("OwnerEntity"));
-
         return tableData;
     }
 
-    public Integer count(String table)
-    {
-        Session session=getSession();
-        String hql="select count(*) from "+table;
-        return ((Long)session.createQuery(hql).uniqueResult()).intValue();
-    }
-
-
-    public TableData getTenantList(TableFilter tableFilter)
-    {
-        Session session = getSession();
-        TableData tableData = new TableData(true);
-        Query query;
-        if (!tableFilter.getSearchValue().equals(""))
-        {
-            String hql = "from TenantEntity t where t.name like (?) or t.phone like (?)";
-            query = session.createQuery(hql).setString(0, "%" + tableFilter.getSearchValue() + "%")
-                    .setString(1, "%" + tableFilter.getSearchValue() + "%");
-        }
-        else
-        {
-            String hql = "from TenantEntity t";
-            query = session.createQuery(hql);
-        }
-        Integer count=query.list().size();
-        List list=query.setFirstResult(tableFilter.getStart()).setMaxResults(tableFilter.getLength()).list();
-
-        tableData.setRecordsFiltered(count);
-        tableData.setJsonString(list);
-        tableData.setRecordsTotal(this.count("TenantEntity"));
-
-        return tableData;
-    }
-
-
-
-    public TableData getAuthenticatedUserList(TableFilter tableFilter)
-    {
-        Session session = getSession();
-        TableData tableData = new TableData(true);
-        Query query;
-        if (!tableFilter.getSearchValue().equals(""))
-        {
-            String hql = "from AuthenticatedUserEntity t where t.name like (?)";
-            query = session.createQuery(hql).setString(0, "%" + tableFilter.getSearchValue() + "%");
-        }
-        else
-        {
-            String hql = "from AuthenticatedUserEntity t";
-            query = session.createQuery(hql);
-        }
-        Integer count=query.list().size();
-        List list=query.setFirstResult(tableFilter.getStart()).setMaxResults(tableFilter.getLength()).list();
-
-        tableData.setRecordsFiltered(count);
-        tableData.setJsonString(list);
-        tableData.setRecordsTotal(this.count("AuthenticatedUserEntity"));
-
-        return tableData;
-    }
 
     @Override
     public TableData getAppUserList(TableFilter tableFilter)
@@ -175,8 +80,6 @@ public class UserDaoImpl implements UserDao
 
         tableData.setRecordsFiltered(count);
         tableData.setJsonString(list);
-        tableData.setRecordsTotal(this.count("AppUserEntity"));
-
         return tableData;
     }
 
@@ -185,11 +88,7 @@ public class UserDaoImpl implements UserDao
     {
         Session session=getSession();
         String hql;
-        if (type== UserType.TENANT)
-            hql="from TenantEntity t where t.phone=:phone";
-        else if (type==UserType.FAMILY)
-            hql="from FamilyEntity t where t.phone=:phone";
-        else if (type==UserType.APPUSER)
+        if (type==UserType.APPUSER)
             hql="from AppUserEntity t where t.phone=:phone";
         else if (type==UserType.OWNER)
             hql="from OwnerEntity t where t.phone=:phone";
@@ -205,37 +104,20 @@ public class UserDaoImpl implements UserDao
     @Override
     public void deleteUserByPhone(String phone, byte type)
     {
+        if (phone==null)
+            return;
         Session session=getSession();
         String hql;
         if (type==UserType.OWNER)
-            hql="delete from OwnerEntity t where t.phone=:phone";
+        {
+            hql = "delete from OwnerEntity t where t.phone=:phone";
+            session.createQuery(hql).setString("phone",phone).executeUpdate();
+        }
         else if (type==UserType.APPUSER)
-            hql="delete from AppUserEntity t where t.phone=:phone";
-        else if (type==UserType.TENANT)
-            hql="delete from TenantEntity t where t.phone=:phone";
-        else if (type==UserType.FAMILY)
-            hql="delete from FamilyEntity t where t.phone=:phone";
-        else
-            return;
-        session.createQuery(hql).setString("phone",phone).executeUpdate();
-    }
-
-    @Override
-    public ArrayList<Object> getOwnersByPropertyID(Integer id)
-    {
-        Session session=getSession();
-        String hql="select o from PropertyOwnerInfoEntity t ,OwnerEntity o where t.propertyId=:id and t.ownerPhone=o.phone";
-        List list=session.createQuery(hql).setInteger("id",id).list();
-        return (ArrayList<Object>) list;
-    }
-
-    @Override
-    public ArrayList<Object> getAppUserByPropertyID(Integer id)
-    {
-        Session session=getSession();
-        String hql="from ";
-        session.createQuery(hql).list();
-        return null;
+        {
+            hql = "update AppUserEntity t set t.status=:status where t.phone=:phone";
+            session.createQuery(hql).setByte("status", AppUserStatus.DELETE).setString("phone",phone).executeUpdate();
+        }
     }
 
     @Override
